refactor: remove UserIsAlreadyVerifiedError from controller errors folder
