--- src/adapters/controllers/verification-token/errors/index.ts
@@ -1,1 +1,0 @@
-export * from '@/adapters/controllers/verification-token/errors/user-is-already-verified-error';
