--- src/adapters/controllers/verification-token/errors/user-is-already-verified-error.ts
@@ -1,6 +1,0 @@
-export class UserIsAlreadyVerifiedError extends Error {
-  constructor() {
-    super('User Is Already Verified');
-    this.name = 'UserIsAlreadyVerifiedError';
-  }
-}
