+++ src/adapters/controllers/verification-token/send-verification-token-to-user/send-verification-token-to-user-controller.ts
@@ -10,7 +10,7 @@ import {
 import { AddVerificationTokenUseCase } from '@/use-cases/verification-token/add-verification-token';
 import { ReadUserUseCase } from '@/use-cases/user/read-user';
 import { SendVerificationTokenUseCase } from '@/use-cases/verification-token/send-verification-token';
-import { UserIsAlreadyVerifiedError } from '@/adapters/controllers/verification-token/errors';
+import { UserIsAlreadyVerifiedError } from '@/use-cases/verification-token/errors';
 
 export class SendVerificationTokenToUserController implements IController {
   constructor(
